--- build.sh
@@ -1,2 +1,0 @@
-#!/bin/sh
-g++ co/*.cpp co/*.S test.cpp -o test -O3 -g -Ico
