+++ co.hpp
@@ -1,0 +1,4 @@
+#include "co/Context.h"
+#include "co/Coroutine.h"
+#include "co/State.h"
+#include "co/Utilities.h"
