+++ co/Context.h
@@ -2,6 +2,7 @@
 #include <cstddef>
 #include <cstring>
 #include <iterator>
+#include "contextswitch.h"
 
 namespace co {
 
@@ -53,5 +54,37 @@ private:
     Word _registers[14];
     char _stack[STACK_SIZE];
 };
+
+
+inline void Context::switchFrom(Context *previous) {
+    contextSwitch(previous, this);
+}
+
+inline void Context::prepare(Context::Callback ret, Context::Word rdi) {
+    Word sp = getSp();
+    fillRegisters(sp, ret, rdi);
+}
+
+inline bool Context::test() {
+    char jojo;
+    ptrdiff_t diff = std::distance(std::begin(_stack), &jojo);
+    return diff >= 0 && diff < STACK_SIZE;
+}
+
+inline Context::Word Context::getSp() {
+    auto sp = std::end(_stack) - sizeof(Word);
+    sp = decltype(sp)(reinterpret_cast<size_t>(sp) & (~0xF));
+    return sp;
+}
+
+inline void Context::fillRegisters(Word sp, Callback ret, Word rdi, ...) {
+    ::memset(_registers, 0, sizeof _registers);
+    auto pRet = (Word*)sp;
+    *pRet = (Word)ret;
+    _registers[RSP] = sp;
+    _registers[RET] = *pRet;
+    _registers[RDI] = rdi;
+}
+
 
 } // co
